Reorganize `exporter/types` into appropriate submodules
