+++ frontend/exporter/src/types/mir.rs
@@ -1,4 +1,9 @@
+//! Copies of the relevant `MIR` types. MIR represents a rust (function) body as a CFG. It's a
+//! semantically rich representation that contains no high-level control-flow operations like loops
+//! or patterns; instead the control flow is entirely described by gotos and switches on integer
+//! values.
 use crate::prelude::*;
+use crate::sinto_as_usize;
 #[cfg(feature = "rustc")]
 use tracing::trace;
 
@@ -1062,11 +1067,167 @@ make_idx_wrapper!(rustc_middle::mir, Local);
 make_idx_wrapper!(rustc_middle::ty, UserTypeAnnotationIndex);
 make_idx_wrapper!(rustc_target::abi, FieldIdx);
 
+/// Reflects [`rustc_middle::mir::UnOp`]
+#[derive_group(Serializers)]
+#[derive(AdtInto, Copy, Clone, Debug, JsonSchema)]
+#[args(<'slt, S: UnderOwnerState<'slt>>, from: rustc_middle::mir::UnOp, state: S as _s)]
+pub enum UnOp {
+    Not,
+    Neg,
+    PtrMetadata,
+}
+
+/// Reflects [`rustc_middle::mir::BinOp`]
+#[derive_group(Serializers)]
+#[derive(AdtInto, Copy, Clone, Debug, JsonSchema)]
+#[args(<'slt, S: UnderOwnerState<'slt>>, from: rustc_middle::mir::BinOp, state: S as _s)]
+pub enum BinOp {
+    // We merge the checked and unchecked variants because in either case overflow is failure.
+    #[custom_arm(
+        rustc_middle::mir::BinOp::Add | rustc_middle::mir::BinOp::AddUnchecked => BinOp::Add,
+    )]
+    Add,
+    #[custom_arm(
+        rustc_middle::mir::BinOp::Sub | rustc_middle::mir::BinOp::SubUnchecked => BinOp::Sub,
+    )]
+    Sub,
+    #[custom_arm(
+        rustc_middle::mir::BinOp::Mul | rustc_middle::mir::BinOp::MulUnchecked => BinOp::Mul,
+    )]
+    Mul,
+    AddWithOverflow,
+    SubWithOverflow,
+    MulWithOverflow,
+    Div,
+    Rem,
+    BitXor,
+    BitAnd,
+    BitOr,
+    #[custom_arm(
+        rustc_middle::mir::BinOp::Shl | rustc_middle::mir::BinOp::ShlUnchecked => BinOp::Shl,
+    )]
+    Shl,
+    #[custom_arm(
+        rustc_middle::mir::BinOp::Shr | rustc_middle::mir::BinOp::ShrUnchecked => BinOp::Shr,
+    )]
+    Shr,
+    Eq,
+    Lt,
+    Le,
+    Ne,
+    Ge,
+    Gt,
+    Cmp,
+    Offset,
+}
+
+/// Reflects [`rustc_middle::mir::ScopeData`]
+#[derive_group(Serializers)]
+#[derive(AdtInto, Clone, Debug, JsonSchema)]
+#[args(<'tcx, S: UnderOwnerState<'tcx> + HasThir<'tcx>>, from: rustc_middle::middle::region::ScopeData, state: S as gstate)]
+pub enum ScopeData {
+    Node,
+    CallSite,
+    Arguments,
+    Destruction,
+    IfThen,
+    Remainder(FirstStatementIndex),
+}
+
+sinto_as_usize!(rustc_middle::middle::region, FirstStatementIndex);
+
+/// Reflects [`rustc_middle::mir::BinOp`]
+#[derive_group(Serializers)]
+#[derive(AdtInto, Clone, Debug, JsonSchema)]
+#[args(<'tcx, S: UnderOwnerState<'tcx> + HasThir<'tcx>>, from: rustc_middle::middle::region::Scope, state: S as gstate)]
+pub struct Scope {
+    pub id: ItemLocalId,
+    pub data: ScopeData,
+}
+
+sinto_as_usize!(rustc_hir::hir_id, ItemLocalId);
+
+#[cfg(feature = "rustc")]
+impl<'tcx, S: UnderOwnerState<'tcx>> SInto<S, ConstantExpr> for rustc_middle::mir::Const<'tcx> {
+    fn sinto(&self, s: &S) -> ConstantExpr {
+        use rustc_middle::mir::Const;
+        let tcx = s.base().tcx;
+        match self {
+            Const::Val(const_value, ty) => {
+                const_value_to_constant_expr(s, *ty, *const_value, rustc_span::DUMMY_SP)
+            }
+            Const::Ty(_ty, c) => c.sinto(s),
+            Const::Unevaluated(ucv, _ty) => {
+                use crate::rustc_middle::query::Key;
+                let span = tcx
+                    .def_ident_span(ucv.def)
+                    .unwrap_or_else(|| ucv.def.default_span(tcx));
+                if ucv.promoted.is_some() {
+                    self.eval_constant(s)
+                        .unwrap_or_else(|| {
+                            supposely_unreachable_fatal!(s, "UnevalPromotedConstant"; {self, ucv});
+                        })
+                        .sinto(s)
+                } else {
+                    match self.translate_uneval(s, ucv.shrink(), span) {
+                        TranslateUnevalRes::EvaluatedConstant(c) => c.sinto(s),
+                        TranslateUnevalRes::GlobalName(c) => c,
+                    }
+                }
+            }
+        }
+    }
+}
+
+#[cfg(feature = "rustc")]
+impl<S> SInto<S, u64> for rustc_middle::mir::interpret::AllocId {
+    fn sinto(&self, _: &S) -> u64 {
+        self.0.get()
+    }
+}
+
+/// Reflects [`rustc_middle::mir::BorrowKind`]
+#[derive(AdtInto)]
+#[args(<S>, from: rustc_middle::mir::BorrowKind, state: S as gstate)]
+#[derive_group(Serializers)]
+#[derive(Copy, Clone, Debug, JsonSchema, Hash, PartialEq, Eq, PartialOrd, Ord)]
+pub enum BorrowKind {
+    Shared,
+    Fake(FakeBorrowKind),
+    Mut { kind: MutBorrowKind },
+}
+
+/// Reflects [`rustc_middle::mir::MutBorrowKind`]
+#[derive(AdtInto)]
+#[args(<S>, from: rustc_middle::mir::MutBorrowKind, state: S as _s)]
+#[derive_group(Serializers)]
+#[derive(Copy, Clone, Debug, JsonSchema, Hash, PartialEq, Eq, PartialOrd, Ord)]
+pub enum MutBorrowKind {
+    Default,
+    TwoPhaseBorrow,
+    ClosureCapture,
+}
+
+/// Reflects [`rustc_middle::mir::FakeBorrowKind`]
+#[derive(AdtInto)]
+#[args(<S>, from: rustc_middle::mir::FakeBorrowKind, state: S as _s)]
+#[derive_group(Serializers)]
+#[derive(Copy, Clone, Debug, JsonSchema, Hash, PartialEq, Eq, PartialOrd, Ord)]
+pub enum FakeBorrowKind {
+    /// A shared (deep) borrow. Data must be immutable and is aliasable.
+    Deep,
+    /// The immediately borrowed place must be immutable, but projections from
+    /// it don't need to be. This is used to prevent match guards from replacing
+    /// the scrutinee. For example, a fake borrow of `a.b` doesn't
+    /// conflict with a mutable borrow of `a.b.c`.
+    Shallow,
+}
+
+sinto_todo!(rustc_ast::ast, InlineAsmTemplatePiece);
+sinto_todo!(rustc_ast::ast, InlineAsmOptions);
 sinto_todo!(rustc_middle::ty, InstanceKind<'tcx>);
 sinto_todo!(rustc_middle::mir, UserTypeProjections);
 sinto_todo!(rustc_middle::mir, LocalInfo<'tcx>);
-sinto_todo!(rustc_ast::ast, InlineAsmTemplatePiece);
-sinto_todo!(rustc_ast::ast, InlineAsmOptions);
 sinto_todo!(rustc_middle::mir, InlineAsmOperand<'tcx>);
 sinto_todo!(rustc_middle::mir, AssertMessage<'tcx>);
 sinto_todo!(rustc_middle::mir, UnwindAction);
@@ -1077,5 +1238,6 @@ sinto_todo!(rustc_middle::mir, MirSource<'tcx>);
 sinto_todo!(rustc_middle::mir, CoroutineInfo<'tcx>);
 sinto_todo!(rustc_middle::mir, VarDebugInfo<'tcx>);
 sinto_todo!(rustc_middle::mir, CallSource);
+sinto_todo!(rustc_middle::mir, UnwindTerminateReason);
 sinto_todo!(rustc_middle::mir::coverage, CoverageKind);
-sinto_todo!(rustc_span, ErrorGuaranteed);
+sinto_todo!(rustc_middle::mir::interpret, ConstAllocation<'a>);
