+++ frontend/exporter/src/types/new/mod.rs
@@ -6,9 +6,11 @@ mod impl_infos;
 mod item_attributes;
 mod predicate_id;
 mod typed_constant_kind;
+mod variant_infos;
 
 pub use full_def::*;
 pub use impl_infos::*;
 pub use item_attributes::*;
 pub use predicate_id::*;
 pub use typed_constant_kind::*;
+pub use variant_infos::*;
