--- frontend/exporter/src/types/todo.rs
@@ -1,21 +1,0 @@
-use crate::prelude::*;
-use crate::sinto_todo;
-sinto_todo!(rustc_middle::ty, ScalarInt);
-sinto_todo!(rustc_middle::ty, AdtFlags);
-sinto_todo!(rustc_middle::ty, NormalizesTo<'tcx>);
-sinto_todo!(rustc_abi, IntegerType);
-sinto_todo!(rustc_abi, ReprFlags);
-sinto_todo!(rustc_abi, Align);
-sinto_todo!(rustc_middle::mir::interpret, ConstAllocation<'a>);
-sinto_todo!(rustc_middle::mir, UnwindTerminateReason);
-sinto_todo!(rustc_ast::tokenstream, DelimSpan);
-sinto_todo!(rustc_ast::tokenstream, DelimSpacing);
-sinto_todo!(rustc_hir::def, DefKind);
-sinto_todo!(rustc_hir, GenericArgs<'a> as HirGenericArgs);
-sinto_todo!(rustc_hir, InlineAsm<'a>);
-sinto_todo!(rustc_target::spec::abi, Abi);
-sinto_todo!(rustc_hir, MissingLifetimeKind);
-sinto_todo!(rustc_hir, QPath<'tcx>);
-sinto_todo!(rustc_hir, WhereRegionPredicate<'tcx>);
-sinto_todo!(rustc_hir, WhereEqPredicate<'tcx>);
-sinto_todo!(rustc_hir, OwnerId);
